feat/adding commission

+++ app/src/main/java/com/paysera/currencyexchanger/di/data/database/entity/TransactionsEntity.kt
@@ -14,4 +14,4 @@ class TransactionsEntity (
     var totalAmount: Double? = null,
 
     @ColumnInfo(name = "count")
-    var count: Double? = null)
+    var count: Int? = null)
